requirements.txt
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
ffmpeg-downloader==0.3.0
ffmpeg-python==0.2.0
future==0.18.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
packaging==23.2
platformdirs==4.1.0
pydub==0.25.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
